requirements.txt
discord.py
youtube-search-python
yt-dlp
pynacl